refactor: UNT-T23489: refactor and documentation update for version 1.4.1
