--- example/global.d.ts
@@ -1,1 +1,0 @@
-declare module 'react-native-spinner-button';
