+++ src/components/AnimatedView/index.ts
@@ -1,2 +1,2 @@
 export { default as AnimatedView } from './AnimatedView';
-export * from './types';
+export * from './AnimatedTypes';
