+++ src/components/ChildrenView/index.ts
@@ -1,2 +1,2 @@
 export { default as ChildrenView } from './ChildrenView';
-export * from './types';
+export * from './ChildrenViewTypes';
