+++ src/components/RadialGradient/index.ts
@@ -1,2 +1,2 @@
 export { default as RadialGradient } from './RadialGradient';
-export * from './types';
+export * from './RadialGradientTypes';
