+++ src/components/SpinnerButton/index.ts
@@ -1,2 +1,2 @@
 export { default as SpinnerButton } from './SpinnerButton';
-export * from './types';
+export * from './SpinnerButtonTypes';
